version 5.2
MAJOR
+ added cross-contamination filter

   After the csf2counts step, the pipeline will execute FILTER_CONTAMINANTS.py
which calls miseq_modules.filter_cross_contaminants().  This function extracts
the MAX consensus sequence from each sample's *.[quality score].conseq file
produced by csf2counts.  This is performed for each quality score cutoff in
[sam2csf_q_cutoffs].  These sequences are written into a temporary file
(*.[dbname].[region].fas) and converted into an index by bowtie2-build.  This
is carried out for each region covered by any of the samples in this run.

  For each region:  All CSF files for this region and given quality score cutoff
are converted into a FASTA format (*.csf.fa).  Each FASTA is mapped against the
region-specific *.bt2 database as an unpaired file by bowtie2-align.  We are not
using the --local option because local alignment should not be necessary at this
stage; this should also make this step much faster.

  The bowtie2-align output is written to *.filtered.sam.  Contents of this file
are split into *.clean.csf and *.contam.csf files.  A sequence is 'clean' if it
maps to the consensus sequence from its own sample.  Otherwise it is considered
to be a 'contaminant'.  NOTE: the *.contam.csf file includes two additional
columns that report the potential contamination source and the read mapping
quality.  When the map quality is low (below 20), it is often the case that the
sequence is very noisy; these should not be considered to be cross-contaminants.
The offsets in the original CSF file were preserved by incorporating these values
into the sequence labels.

  The g2p step and a second csf2counts step are run after this
cross-contamination filter, on the *.clean.csf files.

- removed some unused (block-commented) code in miseqUtils.majority_consensus()

- modified some doc strings in miseqUtils.py  to make them more PEP8-ey

- added function parse_fasta() in miseqUtils.py which is essentially the usual
  convert_fasta() but a little cleaner.

- in miseq_modules.py, made collate_conseqs() and collate_frequencies() more
  robust to variation in the number of '.'-delimited terms in filenames.  This
addresses a problem where *.clean.csf files were causing the pipeline to crash.

- made csf2counts() more robust to very small CSF files (fewer than 5 reads)

- fixed mangled logger.error() in g2p_scoring()