sampleSheetParser now uses a combination of sample name and sample number as keys.  This is more consistent with the way we are generating filenames that are robust to replicate sample names in a given run.
sampleSheetParser now detects the disableContamCheck flag

changed parsing of Description fields in sampleSheetParser to handle
multiplexed regions within a sample, since that is the way they are
annotated - however, the current pipeline does not use this information.
For example, if the user wants to disable contamination checks for nef
but not protease in the same sample, then this can be specified in the
current sample sheet format, but this information is not used in the
pipeline.