Fix #127 by using the full HIV1B-env as the seed for V3LOOP.
Add "Unknown" project.
Don't report a read as an insertion, if part of the insertion is in the
gap between the forward and reverse parts of the read. (Skip lowercase
n's in the InsertionWriter.)
Split login credentials for QAI into project definition and data upload,
so you don't accidentally upload to production after dumping projects.