Add coverage plots back to MISEQ_MONITOR for issue #226.
Also fix some mistakes in the microtest files.
Bring back the instructions for installing R, since we decided to stick
with the R plots under MISEQ_MONITOR.
Stop writing temp files in the /tmp folder, because parallel processes
were writing to the same files.
Uncomment a big section of sample_sheet_parser. Not sure why it was
commented out.