Configure all options through the environment by default
Options such as passwords need to be configured through the
environment. No option must be configured through the command line.
Thus, we are making it simpler by using just the environment in our
example.

Another benefit is that changing these options is easier because
we don't have to re-enable the systemd unit file.