Make bad_cycles.csv argument optional in trim_fastqs.py
Before this change, the file path was required to be provided,
but in case the target file did not exist, the script acted
as if it was empty.

This way optionality was supported, but in a less obvious way.

After this change, the file path is allowed to be None,
which is treated the same as if it does not exist.