Add merge_samples command to micall_docker.py
In this update, a new sub-command `merge_samples` has been added to the `micall_docker.py` file. This enables the combination and filtering of FASTQ files from two samples into a single output file. Various arguments have been provided to specify the input files for this new function, allowing for the specification of both the forward and reverse reads of two samples.
The command combines the reads from both samples, creating a single
FASTQ file. During the process, it also performs censoring.