Contig stitcher: improve cutting of Frankenstein contigs
This solution overcomes the information loss problem of the previous
approach and validates if the cut is permissible - confirming it does
not slice through a SyntheticContig. This function identifies the
segment to be cut, performs the cut, and yields the updated left and
right Frankenstein contigs. Also, a `cut_reference` method was added
to SyntheticContig to indicate that these contigs cannot be cut due to
alignment issues.