Contig stitcher: ensure match_fraction value for every contig
It is required by remap, so we have to provide something to it.
However, it is not clear how to recalculate the value for subparts
of initial contigs since it is initially set by BLAST.

The easiest would be to rerun BLAST on final contigs,
but that also seems wasteful.

Current solution simply copies the match_fraction from parent
of a subpart, or takes min of two "parents" for contigs
that are munged together.