Save git version and history in Singularity container
This commit adds functionality to save the current git version and a
portion of the git history within the Singularity container to improve
traceability and reproducibility.

We capture the git commit hash and the last 10 git commit logs into a
specific directory inside the container. The `.git` directory is
utilized temporarily for this purpose and then removed to keep the
container lightweight.