Split alignments with big deletions in align_consensus
This change tries to revert the behaviour of the aligner
to one that is closer to that of the original mappy-based one.
The original mappy aligner was reporting multiple alignments
if they were far enough apart from each other.
I am not sure how much of that original behaviour is intentional,
or was it just a limitation of mappy.
In case we don't want to split the alignments like that,
we simply need to revert this commit.