Update minimap version to the most recent one
Since minimap base algorithm changed between versions, this update
requires setting the bw and bw_long options to retain the original
behaviour. See lh3/minimap2#1124 for more details.

Also, the quality calculation was adjusted in minimap.
This is not a significant change for us because that quality
metrics is not actually used anywhere.