Do not use deprecated universal lines
In python 3.11, "universal lines" feature [has been deprecated][update].
This feature was implemented as "U" flag that could be passed to:
1) `open(filename, "rU")` type of constructors,
2) `argparse.FileType("rU")` type of constructors.

The flag is no longer supported.

As mentioned in [the update][update], "universal lines" became the
default behaviour. So the flag can be ommited.

This commit removes all the usages of the flag.

[update]: https://docs.python.org/3/whatsnew/3.11.html#porting-to-python-3-11